Split golden replay fixture catalog
